+++ apps/docs/app/foundation/accessibility/page.tsx
@@ -1,0 +1,10 @@
+export default function Accessibility() {
+  return (
+    <div className="layout core">
+      <section>
+        <h1>Accessibility</h1>
+        <p>Work in progress.</p>
+      </section>
+    </div>
+  )
+}